+++ cmd/notifier/main.go
@@ -4,7 +4,6 @@ import (
 	"fmt"
 
 	"github.com/DryginAlexander/notifier/models"
-	"github.com/DryginAlexander/notifier/operator"
 	"github.com/DryginAlexander/notifier/settings"
 	"github.com/DryginAlexander/notifier/web"
 )
@@ -21,8 +20,6 @@ func main() {
 	fmt.Println("applying migration if needed")
 	_ = stor.MigrateDB()
 
-	oper := operator.NewOperator(&stor)
-
-	router := web.Init(&stor, &oper)
+	router := web.Init(&stor)
 	router.Run()
 }
